Revert "Moved test from integration to unit testing"
This reverts commit 473b333f9fbd1b1aaee1af70a67caa2776c664d4.
